+++ services/listgroups/core.go
@@ -25,6 +25,7 @@ import (
 
 	"github.com/BrunoReboul/ram/utilities/aut"
 	"github.com/BrunoReboul/ram/utilities/cai"
+	"github.com/BrunoReboul/ram/utilities/gcf"
 	"github.com/BrunoReboul/ram/utilities/gps"
 	"github.com/BrunoReboul/ram/utilities/ram"
 	"google.golang.org/api/option"
@@ -124,7 +125,7 @@ func Initialize(ctx context.Context, global *Global) {
 // EntryPoint is the function to be executed for each cloud function occurence
 func EntryPoint(ctxEvent context.Context, PubSubMessage gps.PubSubMessage, global *Global) error {
 	// log.Println(string(PubSubMessage.Data))
-	ok, metadata, err := ram.IntialRetryCheck(ctxEvent, global.initFailed, global.retryTimeOutSeconds)
+	ok, metadata, err := gcf.IntialRetryCheck(ctxEvent, global.initFailed, global.retryTimeOutSeconds)
 	if !ok {
 		return err
 	}
@@ -160,8 +161,8 @@ func EntryPoint(ctxEvent context.Context, PubSubMessage gps.PubSubMessage, globa
 }
 
 func initiateQueries(global *Global) error {
-	figures := ram.GetByteSet('0', 10)
-	alphabetLower := ram.GetByteSet('a', 26)
+	figures := ram.getByteSet('0', 10)
+	alphabetLower := ram.getByteSet('a', 26)
 	// Query on directory group email is NOT case sensitive
 	// alphabetUpper := getByteSet('A', 26)
 
@@ -249,9 +250,18 @@ func browseGroups(groups *admin.Groups) error {
 			}
 			publishResult := topic.Publish(ctx, pubSubMessage)
 			waitgroup.Add(1)
-			go ram.GetPublishCallResult(ctx, publishResult, &waitgroup, directoryCustomerID+"/"+group.Email, &pubSubErrNumber, &pubSubMsgNumber, logEventEveryXPubSubMsg)
+			go gps.GetPublishCallResult(ctx, publishResult, &waitgroup, directoryCustomerID+"/"+group.Email, &pubSubErrNumber, &pubSubMsgNumber, logEventEveryXPubSubMsg)
 		}
 	}
 	waitgroup.Wait()
 	return nil
+}
+
+// getByteSet return a set of lenght contiguous bytes starting at bytes
+func getByteSet(start byte, length int) []byte {
+	byteSet := make([]byte, length)
+	for i := range byteSet {
+		byteSet[i] = start + byte(i)
+	}
+	return byteSet
 }
